codes/requirements.txt
fastapi
uvicorn
mysql-connector-python
fastapi
uvicorn
sqlalchemy
pymysql
jinja2
python-multipart
